requirements.txt
streamlit
numpy
faiss-cpu
pypdf
sentence-transformers
rank-bm25
transformers
torch